requirements.txt
PyYAML==5.4.1
schedule==1.1.0
selenium==3.141.0
urllib3==1.26.3
watchdog==2.1.4
yapf==0.30.0
